fix: ajusta logica de criacao e atualizacao de dados manejar o banco de dados adequadamente ao adicionar telefones
