+++ src/index.ts
@@ -15,6 +15,9 @@ app.use(express.json());
 
 app.use("/", router);
 
+app.get("/", (req, res) => res.status(200).json({ test: (typeof req.body.phoneNumbers) }));
+
+
 app.use(ErrorHandler);
 
 app.listen(PORT, () => console.log(`ON: ${PORT}`));
